[legacy] move communication to legacy
